+++ app/src/main/java/opensource/cached_dupe_scanner/MainActivity.kt
@@ -538,8 +538,13 @@ class MainActivity : ComponentActivity() {
                             is Screen.SimilarityClusterDetail -> SimilarityClusterDetailScreen(
                                 repository = similarityRepo,
                                 keepLoadedThumbnailsInMemory = settingsSnapshot.keepLoadedThumbnailsInMemory,
+                                keepLoadedVideoPreviewsInMemory = settingsSnapshot.keepLoadedVideoPreviewsInMemory,
+                                snapVideoPreviewFramesToWidth = settingsSnapshot.snapVideoPreviewFramesToWidth,
+                                videoPreviewLineCount = settingsSnapshot.videoPreviewLineCount,
                                 thumbnailSizeScale = settingsSnapshot.thumbnailSizePercent / 100f,
+                                videoPreviewSizeScale = settingsSnapshot.videoPreviewSizePercent / 100f,
                                 rememberedPreviewCache = rememberedThumbnailCache,
+                                rememberedVideoPreviewCache = rememberedVideoPreviewCache,
                                 showFullPaths = settingsSnapshot.showFullPaths,
                                 deletedPaths = deletedPaths.value,
                                 onDeleteFile = { file ->